+++ lib/dealog_backoffice/messages/aggregates/message.ex
@@ -3,12 +3,25 @@ defmodule DealogBackoffice.Messages.Aggregates.Message do
     :message_id,
     :title,
     :body,
-    :status
+    :status,
+    :rejection_reasons
   ]
 
   alias DealogBackoffice.Messages.Aggregates.Message
-  alias DealogBackoffice.Messages.Commands.{CreateMessage, ChangeMessage, SendMessageForApproval}
-  alias DealogBackoffice.Messages.Events.{MessageCreated, MessageChanged, MessageSentForApproval}
+
+  alias DealogBackoffice.Messages.Commands.{
+    CreateMessage,
+    ChangeMessage,
+    SendMessageForApproval,
+    RejectMessage
+  }
+
+  alias DealogBackoffice.Messages.Events.{
+    MessageCreated,
+    MessageChanged,
+    MessageSentForApproval,
+    MessageRejected
+  }
 
   @doc """
   Create a new message.
@@ -46,6 +59,17 @@ defmodule DealogBackoffice.Messages.Aggregates.Message do
     }
   end
 
+  @doc """
+  Reject an existing message.
+  """
+  def execute(%Message{message_id: message_id}, %RejectMessage{} = reject) do
+    %MessageRejected{
+      message_id: message_id,
+      status: reject.status,
+      reason: reject.reason
+    }
+  end
+
   # State mutators for reconstitution
 
   def apply(%Message{} = message, %MessageCreated{} = created) do
@@ -75,6 +99,15 @@ defmodule DealogBackoffice.Messages.Aggregates.Message do
         title: sent.title,
         body: sent.body,
         status: sent.status
+    }
+  end
+
+  def apply(%Message{} = message, %MessageRejected{} = rejected) do
+    %Message{
+      message
+      | message_id: rejected.message_id,
+        status: rejected.status,
+        rejection_reasons: message.rejection_reasons || [] ++ [rejected.reason]
     }
   end
 end