+++ lib/dealog_backoffice/messages/projectors/message_approval.ex
@@ -4,14 +4,14 @@ defmodule DealogBackoffice.Messages.Projectors.MessageApproval do
     name: "Messages.Projectors.MessageApproval",
     consistency: :strong
 
-  alias DealogBackoffice.Messages.Events.MessageSentForApproval
+  alias DealogBackoffice.Messages.Events.{MessageSentForApproval, MessageRejected}
   alias DealogBackoffice.Messages.Projections.MessageForApproval
   alias DealogBackoffice.Messages
 
   project(%MessageSentForApproval{} = sent, metadata, fn multi ->
     case Messages.get_message_for_approval(sent.message_id) do
-      {:ok, message_approval} ->
-        update_message_approval(multi, message_approval,
+      {:ok, _} ->
+        update_message_approval(multi, sent.message_id,
           title: sent.title,
           body: sent.body,
           status: sent.status,
@@ -23,10 +23,23 @@ defmodule DealogBackoffice.Messages.Projectors.MessageApproval do
     end
   end)
 
-  defp update_message_approval(multi, original_message_approval, changes) do
-    Ecto.Multi.update_all(multi, :update_message_approval, original_message_approval, set: changes)
+  project(%MessageRejected{} = rejected, metadata, fn multi ->
+    update_message_approval(multi, rejected.message_id,
+      status: rejected.status,
+      reason: rejected.reason,
+      updated_at: metadata.created_at
+    )
+  end)
+
+  defp update_message_approval(multi, message_id, changes) do
+    Ecto.Multi.update_all(multi, :update_message_approval, query(message_id), set: changes)
   end
 
+  defp query(message_id) do
+    from(m in MessageForApproval, where: m.id == ^message_id)
+  end
+
+  # Create a new message approval
   defp create_message_approval(multi, sent, metadata) do
     Ecto.Multi.insert(
       multi,