+++ priv/repo/migrations/20200830123512_add_rejection_reason_to_messages_projection.exs
@@ -1,0 +1,9 @@
+defmodule DealogBackoffice.Repo.Migrations.AddRejectionReasonToMessagesProjection do
+  use Ecto.Migration
+
+  def change do
+    alter table(:messages) do
+      add :rejection_reason, :text
+    end
+  end
+end