+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This release has a basic approval process for a single organization on board.
 
+- `ADD` Rejected message can have a reason
+- `ADD` Messages waiting for approval can be rejected
+- `ADD` Dev script has a task to reset the task database
 - `ADD` Messages waiting for approval are listed in the approvals section
 - `ADD` Messages can be sent for approval
 - `ADD` Error messages are displayed as notification